Variation (#2)
This looks less like a mistake, but something comes out at one doesnt it. Yes it does! This is a non-overlookable observation as many things could be simplified if we were to go with the much simpler high level restriction. I didn't have this restriction in mind. Not at all, lucky I found it now. There's going overboard, and then there's "why the hell not?". (I'm leaning towards why the hell not, as a user, and, o man why, when I think that I have to implement it...not that I believe it's difficult (it may not even be difficult, it could be trivial, I doubt it though))